requirements.txt
torch
torchvision
Pillow
scikit-learn
numpy
matplotlib
seaborn
